Change fastify.mjs to server, express is still there for history reasons
